emgm: Drop GTransform
Abstracting that pattern away did not work as hoped, because it would
create ambiguous implicit objects if you wanted more than one transformation
for a subclass of generic.

There does not seem to be a useable work around for this. You can basically
only have one useable subclass of GTransform per Generic subclass. You could
add some userdata parameter and pass that from the wrapper, but this would
get ugly soon, and this is not C.